# api/api_so_r_information_retrieval.go
/*
 * Nsoraf_SOR
 *
 * Nsoraf Steering Of Roaming Service. © 2021, 3GPP Organizational Partners (ARIB, ATIS, CCSA, ETSI, TSDSI, TTA, TTC). All rights reserved.
 *
 * API version: 1.0.1
 * Generated by: OpenAPI Generator (https://openapi-generator.tech)
 */

package api

import (
	"github.com/gin-gonic/gin"
	. "gitlabe2.ext.net.nokia.com/5g_core/sor-af/models"
	"gitlabe2.ext.net.nokia.com/5g_core/sor-af/logger"
	"gitlabe2.ext.net.nokia.com/5g_core/sor-af/producer"
	"net/http"
	"gitlabe2.ext.net.nokia.com/5g_core/sor-af/util"
)

// GetSorInformation - retrieve the steering of roaming information for a UE
func HttpGetSorInformation(c *gin.Context) {
	logger.HandlerLog.Infof("GetSorInformation is called")
	/// <param name="supi type: string">Identifier of the UE</param>
	/// <param name="plmnId type: PlmnId">serving PLMN ID</param>
	/// <param name="supportedFeatures type: string">Supported Features (optional)</param>
	/// <param name="accessType type: AccessType">Access type used by the UE (optional)</param>
	// Getting the path params
	supi := c.Params.ByName("supi")
	// Getting query params
	queryParams := c.Request.URL.Query()
	supportedFeatures := queryParams["supported-features"]
	plmnId := queryParams["plmn-id"]
	accessType := queryParams["access-type"]
	rsp := producer.HandleGetSorInformation(supi, plmnId, supportedFeatures, accessType, c)

	// send response
	for k, v := range rsp.Header {
		// TODO: concatenate all values
		c.Header(k, v[0])
	}
	serializedBody, err := util.Serialize(rsp.Body, "application/json")
	if err != nil {
		logger.HttpLog.Errorf("Serialize Response Body error: %+v", err)
		problemDetail := "Serialize Response Body error: " + err.Error()
		title := "Malformed request syntax"
		var status int32 = http.StatusBadRequest
		rsp := ProblemDetails{
			Title:  &title,
			Status: &status,
			Detail: &problemDetail,
		}
		logger.HttpLog.Errorln(problemDetail)
		c.JSON(http.StatusInternalServerError, rsp)
	} else {
		c.Data(rsp.Status, "application/json", serializedBody)
	}
}