# models/model_access_token_err.go
/*
 * Nsoraf_SOR
 *
 * Nsoraf Steering Of Roaming Service. © 2021, 3GPP Organizational Partners (ARIB, ATIS, CCSA, ETSI, TSDSI, TTA, TTC). All rights reserved.
 *
 * API version: 1.0.1
 * Generated by: OpenAPI Generator (https://openapi-generator.tech)
 */

package datarepository

import ()

// AccessTokenErr - Error returned in the access token response message
type AccessTokenErr struct {
	// simple type

	Error string `json:"error" yaml:"error" bson:"error" mapstructure:"Error"`

	ErrorDescription *string `json:"error_description,omitempty" yaml:"error_description" bson:"error_description" mapstructure:"ErrorDescription"`

	ErrorUri *string `json:"error_uri,omitempty" yaml:"error_uri" bson:"error_uri" mapstructure:"ErrorUri"`
}