# models/model_access_type.go
/*
 * Nsoraf_SOR
 *
 * Nsoraf Steering Of Roaming Service. © 2021, 3GPP Organizational Partners (ARIB, ATIS, CCSA, ETSI, TSDSI, TTA, TTC). All rights reserved.
 *
 * API version: 1.0.1
 * Generated by: OpenAPI Generator (https://openapi-generator.tech)
 */

package datarepository

import ()

type AccessType string

// enum type
// List of AccessType
const (
	ACCESSTYPE__3_GPP_ACCESS    AccessType = "3GPP_ACCESS"
	ACCESSTYPE_NON_3_GPP_ACCESS AccessType = "NON_3GPP_ACCESS"
)