# models/model_redirect_response.go
/*
 * Nsoraf_SOR
 *
 * Nsoraf Steering Of Roaming Service. © 2021, 3GPP Organizational Partners (ARIB, ATIS, CCSA, ETSI, TSDSI, TTA, TTC). All rights reserved.
 *
 * API version: 1.0.1
 * Generated by: OpenAPI Generator (https://openapi-generator.tech)
 */

package datarepository

import ()

type RedirectResponse struct {
	// simple type

	Cause *string `json:"cause,omitempty" yaml:"cause" bson:"cause" mapstructure:"Cause"`

	TargetScp *string `json:"targetScp,omitempty" yaml:"targetScp" bson:"targetScp" mapstructure:"TargetScp"`
}