# models/model_sor_ack_info.go
/*
 * Nsoraf_SOR
 *
 * Nsoraf Steering Of Roaming Service. © 2021, 3GPP Organizational Partners (ARIB, ATIS, CCSA, ETSI, TSDSI, TTA, TTC). All rights reserved.
 *
 * API version: 1.0.1
 * Generated by: OpenAPI Generator (https://openapi-generator.tech)
 */

package datarepository

import (
	"time"
)

type SorAckInfo struct {
	// simple type

	SorAckStatus SorAckStatus `json:"sorAckStatus" yaml:"sorAckStatus" bson:"sorAckStatus" mapstructure:"SorAckStatus"`

	SorSendingTime time.Time `json:"sorSendingTime" yaml:"sorSendingTime" bson:"sorSendingTime" mapstructure:"SorSendingTime"`
}