# models/model_sor_information_privision.go
/*
 * Nsoraf_SOR
 *
 * Nsoraf Steering Of Roaming Service. © 2021, 3GPP Organizational Partners (ARIB, ATIS, CCSA, ETSI, TSDSI, TTA, TTC). All rights reserved.
 *
 * API version: 1.0.1
 * Generated by: OpenAPI Generator (https://openapi-generator.tech)
 */

package datarepository

import ()

type SorInformationPrivision struct {
	// simple type

	// string formatted according to IETF RFC 3986 identifying a referenced resource.

	NotificationDestination string `json:"notificationDestination" yaml:"notificationDestination" bson:"notificationDestination" mapstructure:"NotificationDestination"`

	SorInformation SorInformation `json:"sorInformation" yaml:"sorInformation" bson:"sorInformation" mapstructure:"SorInformation"`
}